add upload files, control button

+++ src/main/java/com/dampcave/courseworkitransitionai/controllers/UserController.java
@@ -1,5 +1,6 @@
 package com.dampcave.courseworkitransitionai.controllers;
 
+import com.dampcave.courseworkitransitionai.models.Role;
 import com.dampcave.courseworkitransitionai.models.User;
 import com.dampcave.courseworkitransitionai.repositoryes.UserRepository;
 import com.dampcave.courseworkitransitionai.service.UserRegistrationRepr;
@@ -20,7 +21,7 @@ import java.util.List;
 
 @Controller
 @RequestMapping("/users")
-@PreAuthorize("hasAuthority('ADMIN')")
+@PreAuthorize("hasAuthority('ADMIN') or hasAuthority('USER')")
 public class UserController {
 
     UserRepository userRepository;
@@ -31,7 +32,7 @@ public class UserController {
     }
 
     @GetMapping()
-    public String showAllUser(Model model){
+    public String showAllUser(Model model) {
         model.addAttribute("users", userRepository.findAll());
         model.addAttribute("title", "Users");
         return "users";
@@ -39,7 +40,7 @@ public class UserController {
 
     @GetMapping("/{user}")
     public String getShowUser(@PathVariable User user,
-                               Model model){
+                              Model model) {
         model.addAttribute("user", user);
         return "user-details";
     }
@@ -47,25 +48,36 @@ public class UserController {
     @PostMapping("/{id}/delete")
     public String deleteUser(@PathVariable(value = "id") Long id, Model model) {
         User user = userRepository.findById(id).orElseThrow();
-        userRepository.delete(user);
+        if (!user.getUsername().equals("admin")) {
+                userRepository.delete(user);
+        }
         return "redirect:/users";
     }
 
     @GetMapping("/deleteall")
-    public String deleteAllUsers() {
-        userRepository.deleteAll();
-        return "redirect:/users";
+    public String deleteAllUsers(HttpServletRequest request, HttpServletResponse response) {
+        List<User> users = userRepository.findAll();
+        for (User user : users) {
+            if (!user.getUsername().equals("admin")) {
+                userRepository.delete(user);
+            }
+        }
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        new SecurityContextLogoutHandler().logout(request, response, auth);
+        return "redirect:/login?logout";
     }
 
     @GetMapping("/blocked")
     public String userBlockedAll(HttpServletRequest request, HttpServletResponse response) {
         List<User> users = userRepository.findAll();
         for (User user : users) {
-            user.setActive(false);
+            if (!user.getUsername().equals("admin")) {
+                user.setActive(false);
+                Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+                new SecurityContextLogoutHandler().logout(request, response, auth);
+            }
             userRepository.save(user);
         }
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        new SecurityContextLogoutHandler().logout(request, response, auth);
         return "redirect:/login?logout";
     }
 
@@ -84,17 +96,17 @@ public class UserController {
                                HttpServletResponse response,
                                @PathVariable(value = "id") Long id) {
         User user = userRepository.findById(id).orElseThrow();
-        user.setActive(!user.isActive());
-        userRepository.save(user);
-
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if (user.getUsername().equals(auth.getName())) {
-            new SecurityContextLogoutHandler().logout(request, response, auth);
-            return "redirect:/login?logout";
+        if (!user.getUsername().equals("admin")) {
+            user.setActive(!user.isActive());
+            userRepository.save(user);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            if (user.getUsername().equals(auth.getName())) {
+                new SecurityContextLogoutHandler().logout(request, response, auth);
+                return "redirect:/login?logout";
+            }
         }
         return "redirect:/users";
     }
-
 
 
 }
